third_party/WebKit/LayoutTests/platform/win/fast/dom/custom/crash-without-dom-wrapper-expected.txt

CONSOLE ERROR: line 1: Uncaught TypeError: Cannot read property 'remove' of null
This is a testharness.js-based test.
PASS crash expected without crbug/446076 fix 
Harness: the test ran to completion.
